refactor: move Ask AI core to shared package
