+++ Packages/MemoraSharedData/Sources/MemoraSharedAskAI/KnowledgeQueryCore.swift
@@ -1,0 +1,693 @@
+import Foundation
+import SwiftData
+import MemoraSharedCore
+import MemoraSharedSchema
+
+@MainActor
+public final class KnowledgeQueryCore {
+
+    // MARK: - Types (referenced from AskAIView — do not move in this task)
+
+    private struct ContextSource {
+        let type: String
+        let title: String
+        let body: String
+    }
+
+    public struct NeutralContextPack: Sendable {
+        public struct Citation: Sendable {
+            public let sourceType: String
+            public let title: String
+            public let excerpt: String
+        }
+
+        public let scopeTitle: String
+        public let promptContext: String
+        public let citations: [Citation]
+        public let instructionHints: [String]
+
+        public init(scopeTitle: String, promptContext: String, citations: [Citation], instructionHints: [String]) {
+            self.scopeTitle = scopeTitle
+            self.promptContext = promptContext
+            self.citations = citations
+            self.instructionHints = instructionHints
+        }
+    }
+
+    // MARK: - Properties
+
+    private let modelContext: ModelContext
+    private let memoryPrivacy: AskAIMemoryPrivacyConfiguration
+    private let retrievalService: AskAIRetrievalService
+
+    // MARK: - Init
+
+    public init(modelContext: ModelContext, memoryPrivacy: AskAIMemoryPrivacyConfiguration) {
+        self.modelContext = modelContext
+        self.memoryPrivacy = memoryPrivacy
+        self.retrievalService = AskAIRetrievalService(modelContext: modelContext, memoryPrivacy: memoryPrivacy)
+    }
+
+    // MARK: - Public API
+
+    public func buildContext(for scope: ChatScope) -> NeutralContextPack {
+        switch scope {
+        case .file(let fileId):
+            return fetchFileContext(fileID: fileId)
+        case .project(let projectId):
+            return fetchProjectContext(projectID: projectId)
+        case .global:
+            return fetchGlobalContext()
+        }
+    }
+
+    /// Query-aware retrieval を利用して context を構築する。
+    /// LocalRetrievalEngine の keyword scoring と rankHint を組み合わせ、
+    /// クエリに関連するチャンクを優先的に取得する。
+    public func buildContext(for scope: ChatScope, query: String) -> NeutralContextPack {
+        let retrievalScope = Self.toRetrievalScope(scope)
+        let retrievalContext = retrievalService.retrieve(scope: retrievalScope, query: query, topN: 8)
+
+        var sources: [ContextSource] = memoryContextSources()
+
+        for retrieved in retrievalContext.chunks {
+            let chunk = retrieved.chunk
+            let scopeName = scopeName(for: scope)
+            if let source = makeContextSource(
+                type: chunk.sourceTypeRaw,
+                title: chunkTitle(sourceType: chunk.sourceType, scopeName: scopeName),
+                body: chunk.text
+            ) {
+                sources.append(source)
+            }
+        }
+
+        // File scope のみ直接コンテキストを追加
+        if case .file(let fileID) = scope, let file = fetchAudioFile(id: fileID) {
+            if sources.isEmpty {
+                // Fallback: direct entity queries when no chunks exist
+                let transcript = retrievalContext.transcriptText
+                let memo = fetchMeetingMemo(for: file.id)?.plainTextCache
+
+                sources += [
+                    makeContextSource(
+                        type: "transcript",
+                        title: "\(file.title) / Transcript",
+                        body: transcript
+                    ),
+                    makeContextSource(
+                        type: "summary",
+                        title: "\(file.title) / Summary",
+                        body: file.summary
+                    ),
+                    makeContextSource(
+                        type: "memo",
+                        title: "\(file.title) / Memo",
+                        body: memo
+                    ),
+                    makeContextSource(
+                        type: "reference",
+                        title: "\(file.title) / Plaud",
+                        body: file.referenceTranscript
+                    )
+                ].compactMap { $0 }
+            }
+        }
+
+        // Project scope fallback
+        if case .project(let projectID) = scope, sources.filter({ $0.type != "memory-profile" && $0.type != "memory-facts" }).isEmpty {
+            let files = fetchAudioFiles(projectID: projectID).prefix(4)
+            let todos = fetchTodos(projectID: projectID).prefix(4)
+
+            for file in files {
+                if let summary = file.summary, !summary.isEmpty {
+                    sources.append(ContextSource(
+                        type: "summary",
+                        title: "\(file.title) / Summary",
+                        body: summary
+                    ))
+                } else if let transcript = fetchTranscript(for: file.id)?.text, !transcript.isEmpty {
+                    sources.append(ContextSource(
+                        type: "transcript",
+                        title: "\(file.title) / Transcript",
+                        body: transcript
+                    ))
+                }
+
+                if let memo = fetchMeetingMemo(for: file.id)?.plainTextCache, !memo.isEmpty {
+                    sources.append(ContextSource(
+                        type: "memo",
+                        title: "\(file.title) / Memo",
+                        body: memo
+                    ))
+                }
+            }
+
+            if !todos.isEmpty {
+                let todoText = todos.map { "• \($0.title)" }.joined(separator: "\n")
+                sources.append(ContextSource(
+                    type: "todo",
+                    title: "\(fetchProject(id: projectID)?.title ?? "Project") / Todo",
+                    body: todoText
+                ))
+            }
+        }
+
+        // Global scope fallback
+        if case .global = scope, sources.filter({ $0.type != "memory-profile" && $0.type != "memory-facts" }).isEmpty {
+            let files = fetchRecentAudioFiles(limit: 5)
+            let todos = fetchTodos(projectID: nil).filter { !$0.isCompleted }.prefix(5)
+
+            for file in files {
+                if let summary = file.summary, !summary.isEmpty {
+                    sources.append(ContextSource(
+                        type: "summary",
+                        title: "\(file.title) / Summary",
+                        body: summary
+                    ))
+                } else if let transcript = fetchTranscript(for: file.id)?.text, !transcript.isEmpty {
+                    sources.append(ContextSource(
+                        type: "transcript",
+                        title: "\(file.title) / Transcript",
+                        body: transcript
+                    ))
+                }
+            }
+
+            if !todos.isEmpty {
+                let todoText = todos.map { "• \($0.title)" }.joined(separator: "\n")
+                sources.append(ContextSource(
+                    type: "todo",
+                    title: "Global / Todo",
+                    body: todoText
+                ))
+            }
+        }
+
+        let scopeTitle = scopeName(for: scope)
+        return makeContextPack(scopeTitle: scopeTitle, sources: sources)
+    }
+
+    // MARK: - Scope Helpers
+
+    private static func toRetrievalScope(_ scope: ChatScope) -> RetrievalScope {
+        switch scope {
+        case .file(let fileId):
+            return .file(fileId)
+        case .project(let projectId):
+            return .project(projectId)
+        case .global:
+            return .global
+        }
+    }
+
+    private func scopeName(for scope: ChatScope) -> String {
+        switch scope {
+        case .file(let fileId):
+            return fetchAudioFile(id: fileId)?.title ?? "このファイル"
+        case .project(let projectId):
+            return fetchProject(id: projectId)?.title ?? "このプロジェクト"
+        case .global:
+            return "Global"
+        }
+    }
+
+    // MARK: - File Scope
+
+    public func fetchFileContext(fileID: UUID) -> NeutralContextPack {
+        guard let file = fetchAudioFile(id: fileID) else {
+            return emptyPack(scopeTitle: "このファイル")
+        }
+
+        var sources: [ContextSource] = memoryContextSources()
+
+        // Primary: KnowledgeChunk-based retrieval (rank descending, top-8)
+        let chunks = fetchChunks(scopeType: .file, scopeID: fileID, limit: 8)
+
+        if !chunks.isEmpty {
+            for chunk in chunks {
+                if let source = makeContextSource(
+                    type: chunk.sourceTypeRaw,
+                    title: chunkTitle(sourceType: chunk.sourceType, scopeName: file.title),
+                    body: chunk.text
+                ) {
+                    sources.append(source)
+                }
+            }
+        } else {
+            // Fallback: direct entity queries when no chunks exist
+            let transcript = fetchTranscript(for: file.id)?.text
+            let memo = fetchMeetingMemo(for: file.id)?.plainTextCache
+
+            sources += [
+                makeContextSource(
+                    type: "transcript",
+                    title: "\(file.title) / Transcript",
+                    body: transcript
+                ),
+                makeContextSource(
+                    type: "summary",
+                    title: "\(file.title) / Summary",
+                    body: file.summary
+                ),
+                makeContextSource(
+                    type: "memo",
+                    title: "\(file.title) / Memo",
+                    body: memo
+                ),
+                makeContextSource(
+                    type: "reference",
+                    title: "\(file.title) / Plaud",
+                    body: file.referenceTranscript
+                )
+            ].compactMap { $0 }
+        }
+
+        return makeContextPack(scopeTitle: file.title, sources: sources)
+    }
+
+    // MARK: - Project Scope
+
+    public func fetchProjectContext(projectID: UUID) -> NeutralContextPack {
+        let project = fetchProject(id: projectID)
+
+        // KnowledgeChunk-based retrieval: rankHint descending, top-6
+        let chunks = fetchChunks(scopeType: .project, scopeID: projectID, limit: 6)
+
+        var sources: [ContextSource] = memoryContextSources()
+
+        for chunk in chunks {
+            if let source = makeContextSource(
+                type: chunk.sourceTypeRaw,
+                title: chunkTitle(sourceType: chunk.sourceType, scopeName: project?.title ?? "Project"),
+                body: chunk.text
+            ) {
+                sources.append(source)
+            }
+        }
+
+        // Fallback: if no chunks found, use direct file queries
+        if chunks.isEmpty {
+            let files = fetchAudioFiles(projectID: projectID).prefix(4)
+            let todos = fetchTodos(projectID: projectID).prefix(4)
+
+            for file in files {
+                if let summary = file.summary, !summary.isEmpty {
+                    sources.append(ContextSource(
+                        type: "summary",
+                        title: "\(file.title) / Summary",
+                        body: summary
+                    ))
+                } else if let transcript = fetchTranscript(for: file.id)?.text, !transcript.isEmpty {
+                    sources.append(ContextSource(
+                        type: "transcript",
+                        title: "\(file.title) / Transcript",
+                        body: transcript
+                    ))
+                }
+
+                if let memo = fetchMeetingMemo(for: file.id)?.plainTextCache, !memo.isEmpty {
+                    sources.append(ContextSource(
+                        type: "memo",
+                        title: "\(file.title) / Memo",
+                        body: memo
+                    ))
+                }
+            }
+
+            if !todos.isEmpty {
+                let todoText = todos.map { "• \($0.title)" }.joined(separator: "\n")
+                sources.append(ContextSource(
+                    type: "todo",
+                    title: "\(project?.title ?? "Project") / Todo",
+                    body: todoText
+                ))
+            }
+        }
+
+        return makeContextPack(scopeTitle: project?.title ?? "このプロジェクト", sources: Array(sources.prefix(6)))
+    }
+
+    // MARK: - Global Scope
+
+    public func fetchGlobalContext() -> NeutralContextPack {
+        // KnowledgeChunk-based retrieval: rankHint + createdAt descending, top-8
+        let chunks = fetchChunks(scopeType: .global, limit: 8)
+
+        var sources: [ContextSource] = memoryContextSources()
+
+        for chunk in chunks {
+            if let source = makeContextSource(
+                type: chunk.sourceTypeRaw,
+                title: chunkTitle(sourceType: chunk.sourceType, scopeName: "Global"),
+                body: chunk.text
+            ) {
+                sources.append(source)
+            }
+        }
+
+        // Fallback: if no chunks found, use direct queries
+        if chunks.isEmpty {
+            let files = fetchRecentAudioFiles(limit: 5)
+            let todos = fetchTodos(projectID: nil).filter { !$0.isCompleted }.prefix(5)
+
+            for file in files {
+                if let summary = file.summary, !summary.isEmpty {
+                    sources.append(ContextSource(
+                        type: "summary",
+                        title: "\(file.title) / Summary",
+                        body: summary
+                    ))
+                } else if let transcript = fetchTranscript(for: file.id)?.text, !transcript.isEmpty {
+                    sources.append(ContextSource(
+                        type: "transcript",
+                        title: "\(file.title) / Transcript",
+                        body: transcript
+                    ))
+                }
+            }
+
+            if !todos.isEmpty {
+                let todoText = todos.map { "• \($0.title)" }.joined(separator: "\n")
+                sources.append(ContextSource(
+                    type: "todo",
+                    title: "Global / Todo",
+                    body: todoText
+                ))
+            }
+        }
+
+        return makeContextPack(scopeTitle: "Memora 全体", sources: Array(sources.prefix(6)))
+    }
+
+    // MARK: - Prompt Building
+
+    public func makePrompt(userMessage: String, contextPack: NeutralContextPack) -> String {
+        let contextBlock = contextPack.promptContext.isEmpty
+            ? "コンテキストはまだありません。一般的な回答ではなく、分からない場合は分からないと答えてください。"
+            : contextPack.promptContext
+        let instructionBlock = contextPack.instructionHints.isEmpty
+            ? ""
+            : "\n追加指示:\n" + contextPack.instructionHints.map { "- \($0)" }.joined(separator: "\n")
+
+        return """
+        あなたは Memora の Ask AI アシスタントです。
+        必ず日本語で簡潔に答えてください。
+        与えられたコンテキストに根拠がある内容だけを優先し、断定できない点は推測だと明示してください。
+        \(instructionBlock)
+
+        スコープ:
+        \(contextPack.scopeTitle)
+
+        コンテキスト:
+        \(contextBlock)
+
+        質問:
+        \(userMessage)
+        """
+    }
+
+    // MARK: - KnowledgeChunk Fetching
+
+    private func fetchChunks(
+        scopeType: KnowledgeChunkScopeType,
+        scopeID: UUID? = nil,
+        sourceType: KnowledgeChunkSourceType? = nil,
+        limit: Int = 6
+    ) -> [KnowledgeChunk] {
+        // Use predicate-based filtering for better performance
+        let scopeTypeRaw = scopeType.rawValue
+
+        var descriptor: FetchDescriptor<KnowledgeChunk>
+        if let scopeID {
+            descriptor = FetchDescriptor<KnowledgeChunk>(
+                predicate: #Predicate {
+                    $0.scopeTypeRaw == scopeTypeRaw && $0.scopeID == scopeID
+                },
+                sortBy: [
+                    SortDescriptor(\.rankHint, order: .reverse),
+                    SortDescriptor(\.createdAt, order: .reverse)
+                ]
+            )
+        } else {
+            descriptor = FetchDescriptor<KnowledgeChunk>(
+                predicate: #Predicate {
+                    $0.scopeTypeRaw == scopeTypeRaw
+                },
+                sortBy: [
+                    SortDescriptor(\.rankHint, order: .reverse),
+                    SortDescriptor(\.createdAt, order: .reverse)
+                ]
+            )
+        }
+        descriptor.fetchLimit = limit * 2
+
+        var results = (try? modelContext.fetch(descriptor)) ?? []
+
+        // Additional sourceType filter if needed (not in predicate to avoid complex predicates)
+        if let sourceType {
+            let sourceTypeRaw = sourceType.rawValue
+            results = results.filter { $0.sourceTypeRaw == sourceTypeRaw }
+        }
+
+        return Array(results.prefix(limit))
+    }
+
+    // MARK: - Memory Context
+
+    private func memoryContextSources() -> [ContextSource] {
+        guard currentMemoryMode != .off else { return [] }
+
+        var sources: [ContextSource] = []
+
+        if let profileSource = makeProfileMemorySource() {
+            sources.append(profileSource)
+        }
+
+        if let factsSource = makeFactsMemorySource() {
+            sources.append(factsSource)
+        }
+
+        return sources
+    }
+
+    private func memoryInstructionHints() -> [String] {
+        guard currentMemoryMode != .off else {
+            return ["memory 設定が完全オフのため、保存済み memory を参照しないでください。"]
+        }
+
+        var hints: [String] = []
+        if currentMemoryMode == .paused {
+            hints.append("新規 memory 保存は停止中ですが、既存の承認済み memory は回答方針に使えます。")
+        }
+        if let profile = fetchPrimaryMemoryProfile() {
+            if let lang = profile.preferredLanguage,
+               !lang.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+                hints.append("preferred language: \(lang)")
+            }
+            if let style = profile.summaryStyle,
+               !style.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+                hints.append("summary style: \(style)")
+            }
+            if let role = profile.roleLabel,
+               !role.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+                hints.append("user role: \(role)")
+            }
+            if let glossary = profile.glossaryJSON,
+               !glossary.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+                hints.append("glossary: \(glossary)")
+            }
+        }
+        return hints
+    }
+
+    private var currentMemoryMode: MemoryPrivacyMode {
+        MemoryPrivacyMode(rawValue: memoryPrivacy.mode) ?? .standard
+    }
+
+    private enum MemoryPrivacyMode: String {
+        case standard
+        case paused
+        case off
+    }
+
+    private func makeProfileMemorySource() -> ContextSource? {
+        guard let profile = fetchPrimaryMemoryProfile() else { return nil }
+
+        let entries = [
+            labeledLine("summaryStyle", value: profile.summaryStyle),
+            labeledLine("preferredLanguage", value: profile.preferredLanguage),
+            labeledLine("roleLabel", value: profile.roleLabel),
+            labeledLine("glossary", value: profile.glossaryJSON)
+        ].compactMap { $0 }
+
+        guard !entries.isEmpty else { return nil }
+
+        return ContextSource(
+            type: "memory-profile",
+            title: "Approved Memory / Profile",
+            body: entries.joined(separator: "\n")
+        )
+    }
+
+    private func makeFactsMemorySource() -> ContextSource? {
+        let activeFacts = fetchActiveMemoryFacts()
+        guard !activeFacts.isEmpty else { return nil }
+
+        let body = activeFacts
+            .prefix(6)
+            .map { fact in
+                let confidence = Int(fact.confidence * 100)
+                return "\(fact.key): \(fact.value) (\(fact.source), \(confidence)%)"
+            }
+            .joined(separator: "\n")
+
+        return ContextSource(
+            type: "memory-facts",
+            title: "Approved Memory / Facts",
+            body: body
+        )
+    }
+
+    // MARK: - Pack Building
+
+    private func makeContextPack(scopeTitle: String, sources: [ContextSource]) -> NeutralContextPack {
+        let limitedSources = sources.prefix(6)
+        let promptContext = limitedSources.map { source in
+            "[\(source.title)]\n\(truncate(source.body, limit: 900))"
+        }.joined(separator: "\n\n")
+
+        let citations = limitedSources.map { source in
+            NeutralContextPack.Citation(
+                sourceType: source.type,
+                title: source.title,
+                excerpt: truncate(source.body, limit: 120)
+            )
+        }
+
+        return NeutralContextPack(
+            scopeTitle: scopeTitle,
+            promptContext: promptContext,
+            citations: citations,
+            instructionHints: memoryInstructionHints()
+        )
+    }
+
+    private func emptyPack(scopeTitle: String) -> NeutralContextPack {
+        NeutralContextPack(
+            scopeTitle: scopeTitle,
+            promptContext: "",
+            citations: [],
+            instructionHints: memoryInstructionHints()
+        )
+    }
+
+    // MARK: - Data Fetching
+
+    private func fetchAudioFile(id: UUID) -> AudioFile? {
+        let descriptor = FetchDescriptor<AudioFile>()
+        return (try? modelContext.fetch(descriptor))?.first(where: { $0.id == id })
+    }
+
+    private func fetchAudioFiles(projectID: UUID) -> [AudioFile] {
+        let descriptor = FetchDescriptor<AudioFile>(
+            sortBy: [SortDescriptor(\.createdAt, order: .reverse)]
+        )
+        return ((try? modelContext.fetch(descriptor)) ?? []).filter { $0.projectID == projectID }
+    }
+
+    private func fetchRecentAudioFiles(limit: Int) -> [AudioFile] {
+        var descriptor = FetchDescriptor<AudioFile>(
+            sortBy: [SortDescriptor(\.createdAt, order: .reverse)]
+        )
+        descriptor.fetchLimit = limit
+        return (try? modelContext.fetch(descriptor)) ?? []
+    }
+
+    private func fetchProject(id: UUID) -> Project? {
+        let descriptor = FetchDescriptor<Project>()
+        return (try? modelContext.fetch(descriptor))?.first(where: { $0.id == id })
+    }
+
+    private func fetchTranscript(for fileID: UUID) -> Transcript? {
+        let descriptor = FetchDescriptor<Transcript>()
+        return (try? modelContext.fetch(descriptor))?.first(where: { $0.audioFileID == fileID })
+    }
+
+    private func fetchMeetingMemo(for fileID: UUID) -> MeetingMemo? {
+        let descriptor = FetchDescriptor<MeetingMemo>()
+        return (try? modelContext.fetch(descriptor))?.first(where: { $0.audioFileID == fileID })
+    }
+
+    private func fetchTodos(projectID: UUID?) -> [TodoItem] {
+        let descriptor = FetchDescriptor<TodoItem>(
+            sortBy: [SortDescriptor(\.createdAt, order: .reverse)]
+        )
+        let todos = (try? modelContext.fetch(descriptor)) ?? []
+        if let projectID {
+            return todos.filter { $0.projectID == projectID }
+        }
+        return todos
+    }
+
+    private func fetchPrimaryMemoryProfile() -> MemoryProfile? {
+        let descriptor = FetchDescriptor<MemoryProfile>(
+            sortBy: [SortDescriptor(\.createdAt, order: .forward)]
+        )
+        return (try? modelContext.fetch(descriptor))?.first
+    }
+
+    private func fetchActiveMemoryFacts() -> [MemoryFact] {
+        guard currentMemoryMode != .off else { return [] }
+
+        let descriptor = FetchDescriptor<MemoryFact>(
+            sortBy: [SortDescriptor(\.confidence, order: .reverse)]
+        )
+
+        return ((try? modelContext.fetch(descriptor)) ?? []).filter { !memoryPrivacy.disabledFactIDs.contains($0.id) }
+    }
+
+    // MARK: - Helpers
+
+    private func sourceLabel(for sourceType: KnowledgeChunkSourceType) -> String {
+        switch sourceType {
+        case .summary: return "Summary"
+        case .transcript: return "Transcript"
+        case .memo: return "Memo"
+        case .todo: return "Todo"
+        case .photoOCR: return "OCR"
+        case .referenceTranscript: return "Reference"
+        }
+    }
+
+    private func chunkTitle(sourceType: KnowledgeChunkSourceType, scopeName: String) -> String {
+        let label = sourceType == .photoOCR ? "Photo OCR" : sourceLabel(for: sourceType)
+        return "\(scopeName) / \(label)"
+    }
+
+    private func makeContextSource(
+        type: String,
+        title: String,
+        body: String?
+    ) -> ContextSource? {
+        guard let body else { return nil }
+        let trimmed = body.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty else { return nil }
+        return ContextSource(
+            type: type,
+            title: title,
+            body: trimmed
+        )
+    }
+
+    private func truncate(_ text: String, limit: Int) -> String {
+        text.count <= limit ? text : String(text.prefix(limit)) + "…"
+    }
+
+    private func labeledLine(_ key: String, value: String?) -> String? {
+        guard let value else { return nil }
+        let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty else { return nil }
+        return "\(key): \(trimmed)"
+    }
+}
